refactor: introduce autogen package (#3316)
* refactor: pass only spec instead of whole policy when possible


* refactor: introduce autogen package
